+++ app/schema.graphql
@@ -198,25 +198,6 @@ type EmbeddingDimension implements Node {
     """HDBSCAN cluster selection epsilon"""
     clusterSelectionEpsilon: Int! = 0
   ): UMAPPoints!
-  hdbscanClustering(
-    """Event ID of the coordinates"""
-    eventIds: [ID!]!
-
-    """Point coordinates. Must be either 2D or 3D."""
-    coordinates2d: [InputCoordinate2D!]
-
-    """Point coordinates. Must be either 2D or 3D."""
-    coordinates3d: [InputCoordinate3D!]
-
-    """HDBSCAN minimum cluster size"""
-    minClusterSize: Int! = 10
-
-    """HDBSCAN minimum samples"""
-    clusterMinSamples: Int! = 1
-
-    """HDBSCAN cluster selection epsilon"""
-    clusterSelectionEpsilon: Int! = 0
-  ): [Cluster!]!
 }
 
 type EmbeddingDimensionConnection {
@@ -357,6 +338,25 @@ type PromptResponse {
 type Query {
   model: Model!
   node(id: GlobalID!): Node!
+  hdbscanClustering(
+    """Event ID of the coordinates"""
+    eventIds: [ID!]!
+
+    """Point coordinates. Must be either 2D or 3D."""
+    coordinates2d: [InputCoordinate2D!]
+
+    """Point coordinates. Must be either 2D or 3D."""
+    coordinates3d: [InputCoordinate3D!]
+
+    """HDBSCAN minimum cluster size"""
+    minClusterSize: Int! = 10
+
+    """HDBSCAN minimum samples"""
+    clusterMinSamples: Int! = 1
+
+    """HDBSCAN cluster selection epsilon"""
+    clusterSelectionEpsilon: Int! = 0
+  ): [Cluster!]!
 }
 
 enum ScalarDriftMetric {
